pyproject.toml
[tool.poetry]
name = "sensorium"
version = "0.1.0"
description = ""
authors = ["Areson"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
drf-spectacular-sidecar = "^2023.10.1"
djoser = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
